project/requirements.txt
certifi==2022.9.24
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
dataiku-api-client==11.0.0
Flask==2.2.2
idna==2.8
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.5.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.21.0
six==1.16.0
urllib3==1.24.3
Werkzeug==2.2.2
zipp==3.8.1
https://design.solutions.dataiku-dss.io/public/packages/dataiku-internal-client.tar.gz